// src/app/dashboard/superadmin/masters/customer/model/customer.model.ts
export interface Customer {
    custAddress: String;
    custCity: String;
    custCode: String;
    custCreatedDate: String;
    custEmail: String;
    custGSTIN: String;
    custId: number;
    custIsActive: boolean;
    custLandline: String;
    custLogoFile: String;
    custMobile: String;
    custName: String;
    custNoOfBranch: number;
    custPincode: String;
    custValidityEnd: String;
    custValidityStart: String;
    refCountryId: number;
    refIndustryTypeId: number;
    refStateId: number;
    stateName: String;
    countryName: String;
}